Add a types module and split some data from the low_level into it.
Add documentation to some of the types that have been moved over.
