+++ wasm-chain-integration/src/v1/trie/foreign.rs
@@ -1,4 +1,4 @@
-use super::low_level::{FlatLoadable, FlatStorable, LoadResult, Reference, StoreResult};
+use super::*;
 
 type LoadCallBack = extern "C" fn(Reference) -> *mut Vec<u8>;
