+++ wasm-chain-integration/src/v1/trie/low_level.rs
@@ -1,145 +1,28 @@
+//! This module provides the low-level primitives that are used to build
+//! the state implementation for V1 smart contracts.
+//!
+//! Functions in this module are, as the name of the module suggests, low-level
+//! and generally have many preconditions, violation of which will make them
+//! unsafe, could trigger panics, or memory corruption. For this reason
+//! functions should only be used via the exposed high-level api in the
+//! [crate::api] module.
+use super::types::*;
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
-use derive_more::{AsRef, From, Into};
 use sha2::Digest;
 use std::{
     collections::HashMap,
-    io::{Read, Seek, SeekFrom, Write},
+    io::Write,
     iter::once,
     slice::Iter,
     sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard},
 };
-use thiserror::*;
 
 const INLINE_CAPACITY: usize = 8;
 
-const INLINE_STEM_LENGTH: usize = 0b0011_1111;
-
-/// A type that can be used to collect auxiliary information while a mutable
-/// trie is being frozen. Particular use-cases of this are collecting the size
-/// of new data, as well as new persistent nodes.
-/// TODO: Methods should probably return Result/Option so that we can terminate
-/// early in case of out of energy.
-pub trait Collector<V> {
-    fn add_value(&mut self, data: &V);
-    fn add_path(&mut self, path: usize);
-    fn add_children(&mut self, num_children: usize);
-}
-/// Collector that does not collect anything.
-pub struct EmptyCollector;
-
-impl<V> Collector<V> for EmptyCollector {
-    #[inline(always)]
-    fn add_value(&mut self, _data: &V) {}
-
-    #[inline(always)]
-    fn add_path(&mut self, _path: usize) {}
-
-    #[inline(always)]
-    fn add_children(&mut self, _num_children: usize) {}
-}
-
-/// A collector that keeps track of how much additional data will be required to
-/// store the tree.
-#[derive(Default)]
-pub struct SizeCollector {
-    num_bytes: u64,
-}
-
-impl SizeCollector {
-    pub fn collect(self) -> u64 { self.num_bytes }
-}
-
-// TODO: Make sure this is adequate. There is a bit of overhead with size length
-// when we store data.
-impl<V: AsRef<[u8]>> Collector<V> for SizeCollector {
-    #[inline]
-    fn add_value(&mut self, data: &V) { self.num_bytes += data.as_ref().len() as u64; }
-
-    #[inline]
-    fn add_path(&mut self, path: usize) {
-        // 1 is for the tag of the value, 4 is for large key length.
-        if path <= INLINE_STEM_LENGTH {
-            self.num_bytes += 1 + path as u64;
-        } else {
-            self.num_bytes += 1 + 4 + path as u64;
-        }
-    }
-
-    fn add_children(&mut self, num_children: usize) {
-        // 1 for the key, 8 for the reference.
-        self.num_bytes += (num_children as u64) * (1 + 8)
-    }
-}
-
-#[repr(transparent)]
-#[derive(Default, Debug, Clone, Copy, Eq, PartialEq, From, Into)]
-/// Reference to a storage location where an item may be retrieved
-pub struct Reference {
-    reference: u64,
-}
-
-#[derive(Debug, Error)]
-pub enum WriteError {
-    #[error("{0}")]
-    IOError(#[from] std::io::Error),
-}
-
-pub type StoreResult<A> = Result<A, WriteError>;
-
-#[derive(Debug, Error)]
-pub enum LoadError {
-    #[error("{0}")]
-    IOError(#[from] std::io::Error),
-    #[error("Incorrect tag")]
-    IncorrectTag {
-        // The tag that was provided.
-        tag: u8,
-    },
-    #[error("Out of bounds read.")]
-    OutOfBoundsRead,
-}
-
-pub type LoadResult<A> = Result<A, LoadError>;
-
-impl Reference {
-    #[inline(always)]
-    pub fn store(&self, sink: &mut impl std::io::Write) -> StoreResult<()> {
-        sink.write_u64::<BigEndian>(self.reference)?;
-        Ok(())
-    }
-}
-
-#[repr(transparent)]
-#[derive(Clone, Copy, AsRef, From, PartialEq, Eq, Ord, PartialOrd)]
-pub struct Hash {
-    hash: [u8; 32],
-}
-
-impl AsRef<[u8]> for Hash {
-    #[inline(always)]
-    fn as_ref(&self) -> &[u8] { self.hash.as_ref() }
-}
-
-impl Hash {
-    #[inline(always)]
-    pub fn zero() -> Self {
-        Self {
-            hash: [0u8; 32],
-        }
-    }
-
-    /// Read a hash value from the provided source, failing if not enough data
-    /// is available.
-    pub fn read(source: &mut impl Read) -> LoadResult<Self> {
-        let mut hash = [0u8; 32];
-        source.read_exact(&mut hash)?;
-        Ok(Self {
-            hash,
-        })
-    }
-}
-
 #[derive(Debug)]
+/// A link to a shared occurrence of a value V.
+/// This is used in this module to construct trees, allowing for sharing of
+/// values in trees and subtrees in case of the persistent tree.
 pub struct Link<V> {
     link: Arc<RwLock<V>>,
 }
@@ -161,12 +44,16 @@ impl<V> Link<V> {
     }
 
     #[inline(always)]
+    /// Immutably borrow the pointed to value.
     pub fn borrow(&self) -> RwLockReadGuard<'_, V> { self.link.as_ref().read().unwrap() }
 
     #[inline(always)]
+    /// Mutably borrow the value that is pointed to.
     pub fn borrow_mut(&self) -> RwLockWriteGuard<'_, V> { self.link.as_ref().write().unwrap() }
 
     #[inline(always)]
+    /// Attempt to consume the link. If the pointed to value has a single owner
+    /// this will return Ok(_), otherwise it will return an error.
     pub fn try_unwrap(self) -> Result<V, Self> {
         Arc::try_unwrap(self.link)
             .map_err(|link| Link {
@@ -176,21 +63,9 @@ impl<V> Link<V> {
     }
 }
 
-pub trait ToSHA256<Ctx> {
-    fn hash(&self, ctx: &mut Ctx) -> Hash;
-}
-
-/// Display the hash in hex.
-impl std::fmt::Debug for Hash {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        for c in self.hash {
-            write!(f, "{:x}", c)?;
-        }
-        Ok(())
-    }
-}
-
 #[derive(Debug, Clone)]
+/// A potentially cached value V. This is a value that can either be purely in
+/// memory, purely in backing storage, or both in memory and in backing storage.
 pub enum CachedRef<V> {
     Disk {
         key: Reference,
@@ -400,8 +275,6 @@ impl<V> CachedRef<V> {
     }
 }
 
-pub type Key = u8;
-
 #[derive(Debug, Clone)]
 enum Stem {
     // the first byte is length, remaining is data.
@@ -474,25 +347,6 @@ impl From<Vec<u8>> for Stem {
     }
 }
 
-// impl FromIterator<u8> for Stem {
-//     #[inline(always)]
-//     fn from_iter<T: IntoIterator<Item = u8>>(iter: T) -> Self where
-//     {
-//         let iter = iter.into_iter();
-//         let len = iter.len();
-//         if len < 15 {
-//             let mut buf = [0u8; 16];
-//             buf[0] = len as u8;
-//             for (place, value) in buf[1..1+len].iter_mut().zip(iter) {
-//                 *place = value;
-//             }
-//             Self::Short(buf)
-//         } else {
-//             Self::Long(Arc::from(iter))
-//         }
-//     }
-// }
-
 impl AsRef<[u8]> for Stem {
     #[inline(always)]
     fn as_ref(&self) -> &[u8] {
@@ -561,53 +415,6 @@ impl<V> Clone for Node<V> {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct Hashed<V> {
-    pub hash: Hash,
-    pub data: V,
-}
-
-impl<V> Hashed<V> {
-    #[inline(always)]
-    pub fn new(hash: Hash, data: V) -> Self {
-        Self {
-            hash,
-            data,
-        }
-    }
-}
-
-impl<Ctx> ToSHA256<Ctx> for Vec<u8> {
-    #[inline(always)]
-    fn hash(&self, _ctx: &mut Ctx) -> Hash {
-        let mut hasher = sha2::Sha256::new();
-        hasher.update(&(self.len() as u64).to_be_bytes());
-        hasher.update(self);
-        let hash = hasher.finalize().into();
-        Hash {
-            hash,
-        }
-    }
-}
-
-impl<Ctx, const N: usize> ToSHA256<Ctx> for [u8; N] {
-    #[inline(always)]
-    fn hash(&self, _ctx: &mut Ctx) -> Hash {
-        let mut hasher = sha2::Sha256::new();
-        hasher.update(&(N as u64).to_be_bytes());
-        hasher.update(self);
-        let hash = hasher.finalize().into();
-        Hash {
-            hash,
-        }
-    }
-}
-
-impl<V, Ctx> ToSHA256<Ctx> for Hashed<V> {
-    #[inline(always)]
-    fn hash(&self, _ctx: &mut Ctx) -> Hash { self.hash }
-}
-
 impl<V: Loadable, Ctx: FlatLoadable> ToSHA256<Ctx> for CachedRef<Hashed<V>>
 where
     V: ToSHA256<Ctx>,
@@ -635,19 +442,8 @@ impl<V, Ctx: FlatLoadable> ToSHA256<Ctx> for Node<V> {
             child_hasher.update(child.1.borrow().hash(ctx));
         }
         hasher.update(child_hasher.finalize());
-        Hash {
-            hash: hasher.finalize().into(),
-        }
-    }
-}
-
-impl<Ctx> ToSHA256<Ctx> for u64 {
-    #[inline(always)]
-    fn hash(&self, _ctx: &mut Ctx) -> Hash {
-        let hash = sha2::Sha256::digest(&self.to_be_bytes()).into();
-        Hash {
-            hash,
-        }
+        let hash: [u8; 32] = hasher.finalize().into();
+        Hash::from(hash)
     }
 }
 
@@ -727,7 +523,7 @@ impl<V> MutableNode<V> {
 /// state rollback. It stores which items were alive at the time of the
 /// checkpoint, utilizing the fact that items are always just added to the end
 /// of the relevant collections.
-pub struct Checkpoint {
+struct Checkpoint {
     pub num_nodes:          usize,
     pub num_values:         usize,
     pub num_borrowed_nodes: usize,
@@ -869,141 +665,6 @@ impl<V> Clone for ChildrenCow<V> {
                 value:      value.clone(),
             },
         }
-    }
-}
-
-pub trait Loadable: Sized {
-    fn load<S: std::io::Read, F: FlatLoadable>(loader: &mut F, source: &mut S) -> LoadResult<Self>;
-
-    fn load_from_location<F: FlatLoadable>(
-        loader: &mut F,
-        location: Reference,
-    ) -> LoadResult<Self> {
-        let mut source = std::io::Cursor::new(loader.load_raw(location)?);
-        Self::load(loader, &mut source)
-    }
-}
-
-/// This loadable instance means that we can only retrieve the vector behind a
-/// cachedref. But it saves on the length which is significant for the concrete
-/// use-case, hence I opted for it.
-impl Loadable for Vec<u8> {
-    fn load<S: std::io::Read, F: FlatLoadable>(
-        _loader: &mut F,
-        source: &mut S,
-    ) -> LoadResult<Self> {
-        let mut ret = Vec::new();
-        source.read_to_end(&mut ret)?;
-        Ok(ret)
-    }
-}
-
-impl<const N: usize> Loadable for [u8; N] {
-    fn load<S: std::io::Read, F: FlatLoadable>(
-        _loader: &mut F,
-        source: &mut S,
-    ) -> LoadResult<Self> {
-        let mut ret = [0u8; N];
-        source.read_exact(&mut ret)?;
-        Ok(ret)
-    }
-}
-
-pub trait FlatStorable {
-    /// Store the provided value and return a reference that can be used
-    /// to load it.
-    fn store_raw(&mut self, data: &[u8]) -> Result<Reference, WriteError>;
-}
-
-pub trait FlatLoadable {
-    type R: AsRef<[u8]>;
-    /// Store the provided value and return a reference that can be used
-    /// to load it.
-    fn load_raw(&mut self, location: Reference) -> LoadResult<Self::R>;
-}
-
-impl FlatStorable for Vec<u8> {
-    fn store_raw(&mut self, data: &[u8]) -> Result<Reference, WriteError> {
-        let len = self.len();
-        let data_len = data.len() as u64;
-        self.extend_from_slice(&data_len.to_be_bytes());
-        self.extend_from_slice(data);
-        Ok((len as u64).into())
-    }
-}
-#[derive(Debug)]
-pub struct Storable<X> {
-    inner: X,
-}
-
-impl<X: Seek + Write> FlatStorable for Storable<X> {
-    fn store_raw(&mut self, data: &[u8]) -> Result<Reference, WriteError> {
-        let pos = self.inner.seek(SeekFrom::Current(0))?;
-        let data_len = data.len() as u32;
-        self.inner.write_u32::<BigEndian>(data_len)?;
-        self.inner.write_all(data)?;
-        Ok(pos.into())
-    }
-}
-
-/// Generic wrapper for a loader. This implements loadable for any S that
-/// implements Seek + Read.
-pub struct Loader<S> {
-    pub inner: S,
-}
-
-impl<S> Loader<S> {
-    pub fn new(file: S) -> Self {
-        Self {
-            inner: file,
-        }
-    }
-}
-
-impl<'a, A: AsRef<[u8]>> FlatLoadable for Loader<A> {
-    type R = Vec<u8>;
-
-    // FIXME: This is inefficient. We allocate too many vectors.
-    fn load_raw(&mut self, location: Reference) -> LoadResult<Self::R> {
-        let slice = self.inner.as_ref();
-        let mut c = std::io::Cursor::new(slice);
-        let pos = c.seek(SeekFrom::Start(location.into()))?;
-        let len = c.read_u64::<BigEndian>()?;
-        let end = (pos + 8 + len) as usize;
-        if end <= slice.len() {
-            Ok(slice[pos as usize + 8..end].to_vec())
-        } else {
-            Err(LoadError::OutOfBoundsRead)
-        }
-    }
-}
-
-impl Loadable for u64 {
-    #[inline(always)]
-    fn load<S: std::io::Read, F: FlatLoadable>(
-        _loader: &mut F,
-        source: &mut S,
-    ) -> LoadResult<Self> {
-        let x = source.read_u64::<BigEndian>()?;
-        Ok(x)
-    }
-}
-impl Loadable for Reference {
-    #[inline(always)]
-    fn load<S: std::io::Read, F: FlatLoadable>(loader: &mut F, source: &mut S) -> LoadResult<Self> {
-        let reference = u64::load(loader, source)?;
-        Ok(reference.into())
-    }
-}
-
-impl<V: Loadable> Loadable for Hashed<V> {
-    fn load<S: std::io::Read, F: FlatLoadable>(loader: &mut F, source: &mut S) -> LoadResult<Self> {
-        let hash = Hash::read(source)?;
-        let data = V::load(loader, source)?;
-        Ok(Hashed {
-            hash,
-            data,
-        })
     }
 }
 
@@ -1191,7 +852,7 @@ fn make_owned<'a, 'b, V>(
     owned_nodes: &'a mut Vec<MutableNode<V>>,
     entries: &'a mut Vec<Entry>,
     loader: &'b mut impl FlatLoadable,
-) -> ((bool, usize), &'a mut tinyvec::TinyVec<[KeyIndexPair; INLINE_CAPACITY]>) {
+) -> (bool, usize, &'a mut tinyvec::TinyVec<[KeyIndexPair; INLINE_CAPACITY]>) {
     let owned_nodes_len = owned_nodes.len();
     let node = unsafe { owned_nodes.get_unchecked(idx) };
     let node_generation = node.generation;
@@ -1252,7 +913,7 @@ fn make_owned<'a, 'b, V>(
         ChildrenCow::Owned {
             value: ref mut children,
             ..
-        } => ((has_value, owned_nodes_len), children),
+        } => (has_value, owned_nodes_len, children),
     }
 }
 
@@ -1281,20 +942,21 @@ type Position = u16;
 pub struct Iterator {
     /// The root of the iterator
     /// This is stored here to allow efficient deleting of iterators.
-    pub(crate) root:         usize,
+    root:         usize,
     /// pointer to the table of nodes.
-    pub(crate) current_node: usize,
+    current_node: usize,
     /// Key at the current position of the iterator.
-    pub(crate) key:          Vec<u8>,
+    key:          Vec<u8>,
     /// Next child to look at. This is None if
     /// we have to give out the value at the current node, and Some(_)
     /// otherwise.
-    pub(crate) next_child:   Option<Position>,
+    next_child:   Option<Position>,
     /// Stack of parents and next positions, and key lengths of parents
-    pub(crate) stack:        Vec<(usize, Position, usize)>,
+    stack:        Vec<(usize, Position, usize)>,
 }
 
 impl Iterator {
+    #[inline(always)]
     pub fn get_key(&self) -> &[u8] { &self.key }
 }
 
@@ -1394,31 +1056,6 @@ impl<V> MutableTrie<V> {
     /// Check whether the current generation is an empty tree.
     pub fn is_empty(&self) -> bool { self.generation_roots.last().map_or(false, |x| x.0.is_none()) }
 }
-
-/// A trait that supports keeping track of resources during tree traversal, to
-/// make sure that resource bounds are not exceeded.
-pub trait TraversalCounter {
-    type Err: std::fmt::Debug;
-    fn tick(&mut self) -> Result<(), Self::Err>;
-}
-/// A counter that does not count anything, and always returns Ok(()).
-pub struct EmptyCounter;
-#[derive(Debug)]
-pub enum NoError {}
-
-impl TraversalCounter for EmptyCounter {
-    type Err = NoError;
-
-    #[inline(always)]
-    fn tick(&mut self) -> Result<(), Self::Err> { Ok(()) }
-}
-
-pub type EntryId = usize;
-
-/// Too many
-#[derive(Debug, Error)]
-#[error("Too many iterators at the same root.")]
-pub struct TooManyIterators;
 
 impl<V> MutableTrie<V> {
     pub fn new_generation(&mut self) {
@@ -1532,7 +1169,7 @@ impl<V> MutableTrie<V> {
                 // we have to visit this child.
                 iterator.stack.push((node_idx, next_child + 1, iterator.key.len()));
                 iterator.next_child = None;
-                let (_, children) =
+                let (_, _, children) =
                     make_owned(node_idx, borrowed_values, owned_nodes, entries, loader);
                 let child = children[usize::from(next_child)];
                 iterator.current_node = child.index();
@@ -1609,7 +1246,7 @@ impl<V> MutableTrie<V> {
                 FollowStem::StemIsPrefix {
                     key_step,
                 } => {
-                    let (_, children) =
+                    let (_, _, children) =
                         make_owned(node_idx, borrowed_values, owned_nodes, entries, loader);
                     let key_usize = usize::from(key_step) << 56;
                     let pair = if let Ok(pair) = children
@@ -1804,7 +1441,7 @@ impl<V> MutableTrie<V> {
                 FollowStem::StemIsPrefix {
                     key_step,
                 } => {
-                    let (_, children) =
+                    let (_, _, children) =
                         make_owned(node_idx, borrowed_values, owned_nodes, entries, loader);
                     let key_usize = usize::from(key_step) << 56;
                     let pair = children
@@ -1844,7 +1481,7 @@ impl<V> MutableTrie<V> {
                         // are automatically invalidated.
                         entries[entry] = Entry::Deleted;
                     }
-                    let (_, children) =
+                    let (_, _, children) =
                         make_owned(node_idx, borrowed_values, owned_nodes, entries, loader);
                     if children.len() == 1 {
                         // collapse path from father
@@ -1880,7 +1517,7 @@ impl<V> MutableTrie<V> {
                         // no children are left, and also no value, we need to delete the child from
                         // the father.
                         if let Some((child_idx, father_idx)) = father {
-                            let ((has_value, _), father_children) = make_owned(
+                            let (has_value, _, father_children) = make_owned(
                                 father_idx,
                                 borrowed_values,
                                 owned_nodes,
@@ -1951,7 +1588,7 @@ impl<V> MutableTrie<V> {
                     if node.locked > 0 {
                         return None;
                     }
-                    let (_, children) =
+                    let (_, _, children) =
                         make_owned(node_idx, borrowed_values, owned_nodes, entries, loader);
                     if let Ok(c_idx) = children.binary_search_by(|ck| ck.key().cmp(&key_step)) {
                         let pair = unsafe { children.get_unchecked(c_idx) };
@@ -1999,7 +1636,7 @@ impl<V> MutableTrie<V> {
                     if node.locked > 0 {
                         return Ok(false);
                     }
-                    let (_, children) =
+                    let (_, _, children) =
                         make_owned(node_idx, borrowed_values, owned_nodes, entries, loader);
                     if let Ok(c_idx) = children.binary_search_by(|ck| ck.key().cmp(&key_step)) {
                         let pair = unsafe { children.get_unchecked(c_idx) };
@@ -2057,7 +1694,7 @@ impl<V> MutableTrie<V> {
                     // fixup the remaining part of the tree. If we deleted the
                     // root the tree is empty, so set it as such.
                     if let Some((child_idx, parent_idx)) = parent_idx {
-                        let ((has_value, _), children) =
+                        let (has_value, _, children) =
                             make_owned(parent_idx, borrowed_values, owned_nodes, entries, loader);
 
                         children.remove(child_idx);
@@ -2234,7 +1871,10 @@ impl<V> MutableTrie<V> {
                     if node.locked > 0 {
                         return None;
                     }
-                    let ((_, owned_nodes_len), children) =
+                    // make_owned may insert additional nodes. Hence we have to update our
+                    // owned_nodes_len to make sure we have the up-to-date
+                    // value.
+                    let (_, owned_nodes_len, children) =
                         make_owned(node_idx, borrowed_values, owned_nodes, entries, loader);
                     let idx = children.binary_search_by(|kk| kk.key().cmp(&key_step));
                     match idx {
